Float to double. Camilo et Thibaut.

+++ src/test/java/tp04/metier/ActionComposeeTest.java
@@ -18,7 +18,7 @@ package tp04.metier;
 import java.util.ArrayList;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
+// import static org.junit.jupiter.api.Assertions.*;
 import java.util.List;
 
 /**
@@ -37,11 +37,11 @@ public class ActionComposeeTest {
         as1 = new ActionSimple("as1");
         as2 = new ActionSimple("as2");
         List<ActionSimple> listeAs = new ArrayList<>();
-        List<Float> listePourcentage = new ArrayList<>();
+        List<Double> listePourcentage = new ArrayList<>();
         listeAs.add(as1);
         listeAs.add(as2);
-        listePourcentage.add(0.4f);
-        listePourcentage.add(0.6f);
+        listePourcentage.add(0.4);
+        listePourcentage.add(0.6);
 
         Assertions.assertTrue(ac.enrgComposition(listeAs, listePourcentage));
     }
@@ -53,13 +53,13 @@ public class ActionComposeeTest {
         as1 = new ActionSimple("as1");
         as2 = new ActionSimple("as2");
         List<ActionSimple> listeAs = new ArrayList<>();
-        List<Float> listePourcentage = new ArrayList<>();
+        List<Double> listePourcentage = new ArrayList<>();
         listeAs.add(as1);
-        listePourcentage.add(0.6f);
+        listePourcentage.add(0.6);
 
         Assertions.assertFalse(ac.enrgComposition(listeAs, listePourcentage));
         listeAs.add(as2);
-        listePourcentage.add(0.6f);
+        listePourcentage.add(0.6);
         Assertions.assertFalse(ac.enrgComposition(listeAs, listePourcentage));
     }
 
@@ -70,11 +70,11 @@ public class ActionComposeeTest {
         as1 = new ActionSimple("as1");
         as2 = new ActionSimple("as2");
         List<ActionSimple> listeAs = new ArrayList<>();
-        List<Float> listePourcentage = new ArrayList<>();
+        List<Double> listePourcentage = new ArrayList<>();
         listeAs.add(as1);
         listeAs.add(as2);
-        listePourcentage.add(1.6f);
-        listePourcentage.add(-0.6f);
+        listePourcentage.add(1.6);
+        listePourcentage.add(-0.6);
 
         Assertions.assertFalse(ac.enrgComposition(listeAs, listePourcentage));
     }
@@ -86,26 +86,27 @@ public class ActionComposeeTest {
         as1 = new ActionSimple("as1");
         as2 = new ActionSimple("as2");
         List<ActionSimple> listeAs = new ArrayList<>();
-        List<Float> listePourcentage = new ArrayList<>();
+        List<Double> listePourcentage = new ArrayList<>();
         listeAs.add(as1);
         listeAs.add(as2);
-        listePourcentage.add(1.6f);
-        listePourcentage.add(-0.6f);
+        listePourcentage.add(1.6);
+        listePourcentage.add(-0.6);
 
         Assertions.assertFalse(ac.enrgComposition(null, listePourcentage));
         Assertions.assertFalse(ac.enrgComposition(listeAs, null));
     }
-        @Test
+
+    @Test
     public void testCreationACListeTailleDifferente() {
         ActionComposee ac = new ActionComposee("ac");
         ActionSimple as1, as2;
         as1 = new ActionSimple("as1");
         as2 = new ActionSimple("as2");
         List<ActionSimple> listeAs = new ArrayList<>();
-        List<Float> listePourcentage = new ArrayList<>();
+        List<Double> listePourcentage = new ArrayList<>();
         listeAs.add(as1);
         listeAs.add(as2);
-        listePourcentage.add(1f);
+        listePourcentage.add(1.);
 
         Assertions.assertFalse(ac.enrgComposition(listeAs, listePourcentage));
     }
